chore:move slice to feature
